requirements.txt
asyncio==3.4.3
click==7.1.2
pathvalidate==2.5.0
Pillow==9.0.1
PyPDF2==1.26.0
requests==2.26.0
requests_html==0.10.0